# openapi/babeltower.v1.yaml
openapi: 3.1.0
info:
  title: BabelTower API
  version: 1.0.0
  description: |
    BabelTower i18n dictionary and translation task API.
    All import and save workflows follow: Input (File) -> Parser -> Standard JSON -> Conflict Check -> Database.

    ## 鉴权
    业务接口通过 HttpOnly Cookie `babeltower_session` 鉴权（HMAC-SHA256 签名的会话 token，默认有效期 8 小时）。
    登录成功后服务端下发该 Cookie，后续请求自动携带。除 `getHealth`、`login`、`logout`、`getCurrentUser`
    之外的接口默认要求登录（`security` 全局生效）；用户管理与系统维护接口额外要求 `ADMIN` 角色（鉴权层面仍是同一 Cookie，
    OpenAPI 无法表达角色，已在各接口描述中标注）。

    ## 统一响应格式
    成功响应统一包裹 `requestId`：`{ ...data, requestId }`。
    错误响应：`{ error: { message, details? }, requestId }`。
    数据库不可用时降级到本地存储，成功响应额外带 `localFallback: true`。
servers:
  - url: /api
    description: Next.js Route Handlers base path
...
security:
  - cookieAuth: []
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Log in with username and password
      description: |
        成功后写入 HttpOnly Cookie `babeltower_session`。
        单会话互踢：每次登录递增 tokenVersion，同账号此前签发的会话全部失效（后登顶前登）。
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  minLength: 1
                password:
                  type: string
                  minLength: 1
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserResponse"
        "400":
          description: Username or password missing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Wrong username or password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Account disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: babeltower_session
      description: HMAC-SHA256 签名的会话 token，登录后由服务端写入 HttpOnly Cookie。
...
  schemas:
    UserRole:
      type: string
      enum: [ADMIN, MAINTAINER]
...
    AuthUser:
      type: object
      required: [id, username, role]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
    AuthUserResponse:
      type: object
      required: [user, requestId]
      properties:
        user:
          $ref: "#/components/schemas/AuthUser"
        localFallback:
          type: boolean
        requestId:
          type: string
...
    ErrorResponse:
      type: object
      required: [error, requestId]
      properties:
        error:
          type: object
          required: [message]
          properties:
            message:
              type: string
            details:
              description: 可选的错误细节，结构因端点而异
        requestId:
          type: string